Refactor extractions.py
